% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{animl}
\alias{animl}
\title{Title}
\usage{
animl(imagedir, mdmodel, speciesmodel, classes)
}
\arguments{
\item{imagedir}{description}

\item{mdmodel}{description}

\item{speciesmodel}{description}

\item{classes}{description}
}
\value{
none
}
\description{
Title
}
\examples{
\dontrun{
imagedir <- "examples/test_data/Southwest"
 mdmodel <- "/mnt/machinelearning/megaDetector/md_v5b.0.0_saved_model"
 modelfile <- "/mnt/machinelearning/Models/Southwest/2022/Southwest_v2.h5"
 classes <- "/mnt/machinelearning/Models/Southwest/2022/classes.txt"
 animl(imagedir,mdmodel,modelfile,classes) }
}
