% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{animl_install}
\alias{animl_install}
\title{Create a miniconda environment for animl and install animl-py}
\usage{
animl_install(
  py_env = "animl_env",
  animl_version = ANIML_VERSION,
  python_version = "3.12",
  confirm = TRUE
)
}
\arguments{
\item{py_env}{name of python environment}

\item{animl_version}{version of animl to install}

\item{python_version}{version of python to install}

\item{confirm}{allow user input}
}
\value{
none
}
\description{
Create a miniconda environment for animl and install animl-py
}
\examples{
\dontrun{animl_install("animl_env", ANIML_VERSION, python_version="3.9", confirm=TRUE)}
}
