% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{animlPipeline}
\alias{animlPipeline}
\title{Run the Full Animl Workflow}
\usage{
animlPipeline(imagedir, workingdir, mdmodel, classmodel, classlabels)
}
\arguments{
\item{imagedir}{the directory path containing images to analyze}

\item{workingdir}{the directory path to save outputs to}

\item{mdmodel}{location of the MegaDetector model}
}
\value{
a file manifest with predictions
}
\description{
Run the Full Animl Workflow
}
\examples{
\dontrun{
animlPipeline("~/examples/test_data/", "home/kyra/test/", NULL, NULL, NULL)
}
}
