% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyPredictions.R
\name{applyPredictions}
\alias{applyPredictions}
\title{Apply Classifier Predictions and Merge DataFrames}
\usage{
applyPredictions(animals, pred, classfile, outfile = NULL, counts = FALSE)
}
\arguments{
\item{animals}{Set of animal crops/images}

\item{pred}{Classifier predictions for animal crops/images}

\item{classfile}{.txt file containing common names for species classes}

\item{outfile}{File to which results are saved}

\item{counts}{Returns a table of all predictions, defaults to FALSE}
}
\value{
fully merged dataframe with Species predictions and confidence weighted by MD conf
}
\description{
Apply Classifier Predictions and Merge DataFrames
}
\examples{
\dontrun{
alldata <- applyPredictions(animals,empty,classfile,pred,counts = FALSE)
}
}
