% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDirectories.R
\name{build_Directories}
\alias{build_Directories}
\title{Initialize Directories}
\usage{
build_Directories(basedir)
}
\arguments{
\item{basedir}{The directory in which all data and temporary files will be saved.}
}
\value{

}
\description{
Initialize Directories
}
