% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{build_file_manifest}
\alias{build_file_manifest}
\title{File Management Module}
\usage{
build_file_manifest(
  image_dir,
  exif = TRUE,
  out_file = NULL,
  offset = 0,
  recursive = TRUE
)
}
\arguments{
\item{image_dir}{folder to search through and find media files}

\item{exif}{returns date and time information from exif data, defaults to true}

\item{out_file}{directory to save .csv of manifest to}

\item{offset}{add offset in hours for videos when using the File Modified date, defaults to 0}

\item{recursive}{Should directories be scanned recursively? Default TRUE}
}
\value{
files dataframe with or without file dates
}
\description{
This module provides functions and classes for managing files and directories.
}
\details{
Kyra Swanson 2023
Find Image/Video Files and Gather exif Data
}
\examples{
\dontrun{
files <- build_file_manifest("C:\\\\Users\\\\usr\\\\Pictures\\\\")
}
}
