% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileManifest.R
\name{buildFileManifest}
\alias{buildFileManifest}
\title{Extract exif data and create dataframe, adjust time if necessary.}
\usage{
buildFileManifest(imagedir, outfile = NULL, timezone_offset = 0)
}
\arguments{
\item{imagedir}{file path}

\item{outfile}{file path to which the data frame should be saved}

\item{timezone_offset}{integer to adjust file modify time}
}
\value{
images
}
\description{
Extract exif data and create dataframe, adjust time if necessary.
}
\examples{
\dontrun{
images <- extractFiles("C:\\\\Users\\\\usr\\\\Pictures\\\\")
}
}
