% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{check_animl_py}
\alias{check_animl_py}
\title{Check if animl-py can connect to exiftool and CUDA}
\usage{
check_animl_py()
}
\value{
A list with three logical values:
\item{exiftool}{TRUE if exiftool is available}
\item{torch_cuda}{TRUE if CUDA is available to PyTorch}
\item{onnx_cuda}{TRUE if CUDA is available to ONNX}
}
\description{
Check if animl-py can connect to exiftool and CUDA
}
\examples{
\dontrun{
  result <- check_animl_py()
  print(result)
}

}
