% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManagement.R
\name{checkFile}
\alias{checkFile}
\title{Check for files existence and prompt user if they want to load}
\usage{
checkFile(file)
}
\arguments{
\item{file}{the full path of the file to check}
}
\value{
a boolean indicating wether a file was found
and the user wants to load or not
}
\description{
Check for files existence and prompt user if they want to load
}
\examples{
\dontrun{
  checkFile("path/to/newfile.csv")
}
}
