% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyMD.R
\name{classifyImagesBatchMD_0ld}
\alias{classifyImagesBatchMD_0ld}
\title{Run Megadetector on a batch of images (old)}
\usage{
classifyImagesBatchMD_0ld(mdsession, images)
}
\arguments{
\item{mdsession}{should be the output from loadMDmodel(model)}

\item{images}{a vector containing image filepaths}
}
\value{
a list of lists of bounding boxes for each image
}
\description{
Old version for applying MD classification, does not use parallel processing.
}
