% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifySpecies.R
\name{classifySpecies}
\alias{classifySpecies}
\title{Classifies Crops Using Specified Models}
\usage{
classifySpecies(
  mdresults,
  model,
  resize = 299,
  standardize = TRUE,
  batch_size = 32,
  workers = 1
)
}
\arguments{
\item{mdresults}{flattened mdresults dataframe}

\item{model}{models with which to classify species}

\item{resize}{resize images before classification, defaults to 299x299px}

\item{standardize}{standardize images, defaults to FALSE}

\item{batch_size}{number of images processed in each batch (keep small)}

\item{workers}{number of cores}
}
\value{
a matrix of likelihoods for each class for each image
}
\description{
Classifies Crops Using Specified Models
}
\examples{
\dontrun{
pred<-classifySpecies(imagesallanimal,paste0(modelfile,".h5"),
                      resize=456,standardize=FALSE,batch_size = 64,workers=8)
}
}
