% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifySequence.R
\name{classifyVideo}
\alias{classifyVideo}
\title{Select Best Classification From Multiple Frames}
\usage{
classifyVideo(mdanimals, how = "count")
}
\arguments{
\item{mdanimals}{dataframe of all frames including species classification}

\item{how}{method for selecting best prediction, defaults to most frequent}
}
\value{
dataframe with new prediction in "Common" column
}
\description{
Select Best Classification From Multiple Frames
}
\examples{
\dontrun{
mdanimals <- classifyVideo(mdanimals)
}
}
