% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reid.R
\name{compute_distance_matrix}
\alias{compute_distance_matrix}
\title{A wrapper function for computing distance matrix.}
\usage{
compute_distance_matrix(input1, input2, metric = "euclidean")
}
\arguments{
\item{input1}{2-D feature matrix}

\item{input2}{2-D feature matrix}

\item{metric}{"euclidean" or "cosine", Default is "euclidean"}
}
\value{
distance matrix
}
\description{
A wrapper function for computing distance matrix.
}
\examples{
\dontrun{dist_matrix <- compute_distance_matrix(embeddings, embeddings, metric='cosine')}
}
