% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMDresults.R
\name{convertMDResultsSimple}
\alias{convertMDResultsSimple}
\title{Quickly flatten output from MegaDetector}
\usage{
convertMDResultsSimple(mdresults)
}
\arguments{
\item{mdresults}{list of bounding boxes for each image}
}
\value{
A dataframe with one entry for each bounding box
}
\description{
Returns a dataframe where each row is a MD bounding box, does not take into consideration
potential collisions with the outer edge of the image.
}
\examples{

convertMDResultsSimple(mdresults)

}
