% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctResults.R
\name{correctResults}
\alias{correctResults}
\title{Title}
\usage{
correctResults(resultsfile, linkdir)
}
\arguments{
\item{resultsfile}{final results file with predictions, expects a "UniqueName" column}

\item{linkdir}{symlink directory that has been validated}
}
\value{
dataframe with new "Species" column that contains the verifed species
}
\description{
Title
}
\examples{
\dontrun{}
}
