% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reid.R
\name{cosine_distance}
\alias{cosine_distance}
\title{Computes cosine distance of two sets of vectors}
\usage{
cosine_distance(input1, input2)
}
\arguments{
\item{input1}{2-D feature matrix}

\item{input2}{2-D feature matrix}
}
\value{
distance matrix
}
\description{
Computes cosine distance of two sets of vectors
}
\examples{
\dontrun{dist_matrix <- cosine_distance(embeddings, embeddings)}
}
