% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{create_pyenv}
\alias{create_pyenv}
\title{Create the environment animl_env}
\usage{
create_pyenv(python_version, py_env = "animl_env")
}
\arguments{
\item{python_version}{python version for new environment}

\item{py_env}{name of python environment}
}
\value{
python path for new environment
}
\description{
Create the environment animl_env
}
\examples{
\dontrun{create_env("3.12", py_env='animl_env')}
}
