% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxes.R
\name{crop_images}
\alias{crop_images}
\title{Crops all images from an input file with specific required columns: Frame,
bbox1, bbox2, bbox3, and bbox4}
\usage{
crop_images(imagelist, outdir)
}
\arguments{
\item{imagelist}{The path for the input csv file}

\item{outdir}{The path where generated cropped images should be uploaded}
}
\value{
no return value, outputs the cropped image
}
\description{
Crops all images from an input file with specific required columns: Frame,
bbox1, bbox2, bbox3, and bbox4
}
\examples{
\dontrun{
cropImagesFromFile("/image/path/file.csv", "/output/path/")
}
}
