% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decodeImage.R
\name{decode_img_resize}
\alias{decode_img_resize}
\title{Load image with TF 35722 and resize for MD}
\usage{
decode_img_resize(image_filepath, height = 299, width = 299)
}
\arguments{
\item{image_filepath}{the filepath of the image}

\item{height}{resize height, defaults to 299 px}

\item{width}{resize width, defualts to 299 px}
}
\description{
Load image with TF 35722 and resize for MD
}
\examples{
\dontrun{
  dataset<-tfdatasets::dataset_map_and_batch(dataset,decode_img_resize,batch_size)
}
}
