% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{detectMD_batch}
\alias{detectMD_batch}
\title{Apply MegaDetector to a Given Batch of Images}
\usage{
detectMD_batch(
  detector,
  image_file_names,
  checkpoint_path = NULL,
  checkpoint_frequency = -1,
  confidence_threshold = 0.1,
  quiet = TRUE,
  image_size = NULL,
  file_col = "Frame"
)
}
\arguments{
\item{detector}{preloaded md model}

\item{image_file_names}{list of image filenames, a single image filename, or folder}

\item{checkpoint_path}{path to checkpoint file}

\item{checkpoint_frequency}{write results to checkpoint file every N images}

\item{confidence_threshold}{only detections above this threshold are returned}

\item{quiet}{print debugging statements when false, defaults to true}

\item{image_size}{overrides default image size, 1280}

\item{file_col}{select which column if image_file_names is a manifest}
}
\value{
list of dictionaries of MegaDetector detections
}
\description{
Apply MegaDetector to a Given Batch of Images
}
\examples{
\dontrun{mdres <- detectMD_batch(md_py, allframes$Frame)}
}
