% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_model}
\alias{download_model}
\title{Download a given model}
\usage{
download_model(model_url, out_dir = "models")
}
\arguments{
\item{model_url}{url of the model to download}

\item{out_dir}{Directory to save the model.}
}
\value{
None
}
\description{
Download a given model
}
\examples{
\dontrun{download_model('animl_py$CLASSIFIER$get('SDZWA_savanna_v3')')}
}
