% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reid.R
\name{euclidean_squared_distance}
\alias{euclidean_squared_distance}
\title{Computes euclidean squared distance of two sets of vectors}
\usage{
euclidean_squared_distance(input1, input2)
}
\arguments{
\item{input1}{2-D feature matrix}

\item{input2}{2-D feature matrix}
}
\value{
distance matrix
}
\description{
Computes euclidean squared distance of two sets of vectors
}
\examples{
\dontrun{dist_matrix <- euclidean_squared_distance(embeddings, embeddings)}
}
