% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_folders}
\alias{export_folders}
\title{Create SymLink Directories and Sort Classified Images}
\usage{
export_folders(
  manifest,
  out_dir,
  out_file,
  label_col = "prediction",
  file_col = "filepath",
  unique_name = "uniquename",
  copy = FALSE
)
}
\arguments{
\item{manifest}{DataFrame of classified images}

\item{out_dir}{Destination directory for symlinks}

\item{out_file}{}

\item{file_col}{Colun containing file paths}

\item{unique_name}{Unique image name identifier}

\item{copy}{Toggle to determine copy or hard link, defaults to link}
}
\value{
manifest with added link columns
}
\description{
Create SymLink Directories and Sort Classified Images
}
\examples{
\dontrun{
manifest <- sort_species(manifest, link_dir)
}
}
