% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_timelapse}
\alias{export_timelapse}
\title{Title}
\usage{
export_timelapse(animals, empty, imagedir, only_animal = TRUE)
}
\arguments{
\item{animals}{a DataFrame that has entries of anuimal classification}

\item{empty}{a DataFrame that has detection of non-animal objects in images}

\item{imagedir}{location of root directory where all images are stored (can contain subdirectories)}

\item{only_animal}{A bool that confirms whether we want only animal detctions or all}
}
\value{
animals.csv, non-anim.csv, csv_loc
}
\description{
Title
}
\examples{
\dontrun{export_timelapse(animals, empty, '/path/to/images/')}
}
