% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reid.R
\name{extract_embeddings}
\alias{extract_embeddings}
\title{Extract Embeddings from MiewID}
\usage{
extract_embeddings(
  manifest,
  miew_model,
  file_col = "FilePath",
  batch_size = 1,
  workers = 1
)
}
\arguments{
\item{manifest}{list of files}

\item{miew_model}{loaded miewid model}

\item{file_col}{column name containing file paths}

\item{batch_size}{batch size for generator}

\item{workers}{number of workers for generator}
}
\value{
matrix of embeddings
}
\description{
Extract Embeddings from MiewID
}
\examples{
\dontrun{embeddings = extract_embeddings(manifest, miew)}
}
