% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{extract_major_minor}
\alias{extract_major_minor}
\title{Compare to Animl-Py version, ignoring patch number}
\usage{
extract_major_minor(version_string)
}
\arguments{
\item{version_string}{the version to extract substring for}
}
\value{
version_string with only major and minor release
}
\description{
Compare to Animl-Py version, ignoring patch number
}
\examples{
\dontrun{extract_major_minor("3.4.1")}
}
