% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reid.R
\name{extract_miew_embeddings}
\alias{extract_miew_embeddings}
\title{Extract Embeddings from MiewID}
\usage{
extract_miew_embeddings(
  miew_model,
  manifest,
  file_col = "filepath",
  batch_size = 1,
  num_workers = 1,
  device = NULL
)
}
\arguments{
\item{miew_model}{loaded miewid model}

\item{manifest}{list of files}

\item{file_col}{column name containing file paths}

\item{batch_size}{batch size for generator}

\item{num_workers}{number of workers for generator}

\item{device}{device to run model on}
}
\value{
matrix of embeddings
}
\description{
Extract Embeddings from MiewID
}
\examples{
\dontrun{embeddings = extract_embeddings(manifest, miew)}
}
