% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBoxes.R
\name{extractAllBoxes}
\alias{extractAllBoxes}
\title{Extract bounding boxes and save as new image from a batch of images}
\usage{
extractAllBoxes(
  images,
  min_conf = 0,
  buffer = 2,
  save = F,
  resize = NA,
  outdir = "",
  quality = 0.8,
  parallel = F,
  nproc = parallel::detectCores()
)
}
\arguments{
\item{images}{list of images, raw MD output}

\item{min_conf}{Confidence threshold (defaults to 0, not in use)}

\item{buffer}{Adds a buffer to the MD bounding box, defaults to 2px}

\item{save}{T/F toggle to save output cropped, defaults to False}

\item{resize}{Size in pixels to resize cropped images, NA if images are not resized, defaults to NA}

\item{outdir}{Directory in which output cropped images will be saved}

\item{quality}{Compression level of output cropped image, defaults to 0.8}

\item{parallel}{T/F toggle to enable parallel processing, defaults to False}

\item{nproc}{Number of workers if parallel = True, defaults to output of detectCores()}
}
\value{
a flattened dataframe containing crop information
}
\description{
Extract bounding boxes and save as new image from a batch of images
}
