% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBoxes.R
\name{extractAllBoxesFromFlat}
\alias{extractAllBoxesFromFlat}
\title{Extract ,cropped images from a processed dataframe}
\usage{
extractAllBoxesFromFlat(
  images,
  buffer = 0,
  resize = NA,
  quality = 0.8,
  outdir = "",
  save = F,
  overwrite = T,
  parallel = F,
  nproc = detectCores()
)
}
\arguments{
\item{buffer}{Adds a buffer to the MD bounding box, defaults to 2px}

\item{resize}{Size in pixels to resize cropped images, NA if images are not resized, defaults to NA}

\item{quality}{Compression level of output cropped image, defaults to 0.8}

\item{outdir}{Directory in which output cropped images will be saved}

\item{save}{T/F toggle to save output cropped, defaults to False}

\item{overwrite}{T/F toggle to overwrite output cropped images if they already exis, defaults to True}

\item{parallel}{T/F toggle to enable parallel processing, defaults to False}

\item{nproc}{Number of workers if parallel = True, defaults to output of detectCores()}

\item{image}{dataframe containing MD output (assumes single row)}

\item{plot}{T/F toggle to plot each crop in the plot window, defaults to True}
}
\value{
A dataframe containing image and crop paths,
}
\description{
Extract ,cropped images from a processed dataframe
}
\examples{

extractBoxes(mdresultsflat)
}
