% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBoxes.R
\name{extractBoxesFromFlat}
\alias{extractBoxesFromFlat}
\title{Extract crops from a single image represented by a processed dataframe}
\usage{
extractBoxesFromFlat(
  image,
  min_conf = 0,
  buffer = 0,
  plot = T,
  save = F,
  resize = NA,
  outdir = "",
  quality = 0.8
)
}
\arguments{
\item{image}{dataframe containing MD output (assumes single row)}

\item{min_conf}{Confidence threshold (defaults to 0, not in use)}

\item{buffer}{Adds a buffer to the MD bounding box, defaults to 2px}

\item{plot}{T/F toggle to plot each crop in the plot window, defaults to True}

\item{save}{T/F toggle to save output cropped, defaults to False}

\item{resize}{Size in pixels to resize cropped images, NA if images are not resized, defaults to NA}

\item{outdir}{Directory in which output cropped images will be saved}

\item{quality}{Compression level of output cropped image, defaults to 0.8}
}
\value{

}
\description{
Extract crops from a single image represented by a processed dataframe
}
\examples{

extractBoxes(mdresultsflat)
}
