% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videoProcessing.R
\name{extractFrames}
\alias{extractFrames}
\title{Extract frames from video for classification}
\usage{
extractFrames(
  files,
  outdir = tempfile(),
  outfile = NULL,
  fps = NULL,
  frames = NULL,
  parallel = FALSE,
  workers = 1,
  checkpoint = 1000
)
}
\arguments{
\item{files}{dataframe of videos}

\item{outdir}{directory to save frames to}

\item{outfile}{file to which results will be saved}

\item{fps}{frames per second, otherwise determine mathematically}

\item{frames}{number of frames to sample}

\item{parallel}{Toggle for parallel processing, defaults to FALSE}

\item{workers}{number of processors to use if parallel, defaults to 1}

\item{checkpoint}{if not parallel, checkpoint ever n files, defaults to 1000}
}
\value{
dataframe of still frames for each video
}
\description{
This function can take
}
\examples{
\dontrun{
frames <- extractFrames(videos, outdir = "C:\\\\Users\\\\usr\\\\Videos\\\\", frames = 5)
}
}
