% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videoProcessing.R
\name{extractFramesSingle}
\alias{extractFramesSingle}
\title{Extract Frames for Single Video}
\usage{
extractFramesSingle(x, outdir, fps = NULL, frames = NULL)
}
\arguments{
\item{x}{filepath to image}

\item{outdir}{directory to save frames to}

\item{fps}{number of frames per second to save}

\item{frames}{number of frames evenly distributed to save}
}
\value{
dataframe of filepaths, frame paths
}
\description{
Extract Frames for Single Video
}
\examples{
\dontrun{
result <- extractFramesSingle(video$FilePath, outdir, frames=3)
}
}
