% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{get_animals}
\alias{get_animals}
\title{Return a dataframe of only MD animals}
\usage{
get_animals(manifest)
}
\arguments{
\item{manifest}{all megadetector frames}
}
\value{
animal frames classified by MD
}
\description{
Return a dataframe of only MD animals
}
\examples{
\dontrun{
animals <- get_animals(imagesall)
}
}
