% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitData.R
\name{getEmpty}
\alias{getEmpty}
\title{Return MD empty, vehicle and human images in a dataframe}
\usage{
getEmpty(manifest)
}
\arguments{
\item{manifest}{all megadetector frames}
}
\value{
list of empty/human/vehicle allframes with md classification
}
\description{
Return MD empty, vehicle and human images in a dataframe
}
\examples{
\dontrun{
empty <- getEmpty(imagesall)
}
}
