% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{imageAugmentationGeometry}
\alias{imageAugmentationGeometry}
\title{Perform random geometric transformations on an image.}
\usage{
imageAugmentationGeometry()
}
\value{
A keras model.
}
\description{
Returns a keras model that performs random geometric transformations on an image.
}
\examples{

}
