% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{imageLabelCrop}
\alias{imageLabelCrop}
\title{Load image, crop and return a tensor with an image and a corresponding label.}
\usage{
imageLabelCrop(data, classes, height = 299, width = 299, standardize = FALSE)
}
\arguments{
\item{data}{a list with the first element being an image file path, the next four elements being the bounding box coordinates and the last element a label}

\item{height}{the height the cropped image will be resized to.}

\item{width}{the width the cropped image will be resized to.}

\item{standardize}{standardize the image, TRUE or FALSE.}
}
\value{
An image and label tensor.
}
\description{
Load image, crop and return a tensor with an image and a corresponding label. Internal function to be called by image generator function.
}
\examples{
\dontrun{}
}
