% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{list_models}
\alias{list_models}
\title{List available models for download.}
\usage{
list_models()
}
\value{
printout of models
}
\description{
List available models for download.
}
\examples{
\dontrun{
  list_models()
  download_model("https://models.com/path/to/model.pt", out_dir='models')
}
}
