% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{load_detector}
\alias{load_detector}
\title{Load an Object Detector}
\usage{
load_detector(model_path, model_type, device = NULL)
}
\arguments{
\item{model_path}{path to detector model file}

\item{model_type}{type of model expected \link{"MDV5", "MDV6", "YOLO"}}

\item{device}{specify to run on cpu or gpu}
}
\value{
megadetector object
}
\description{
Load an Object Detector
}
\examples{
\dontrun{md_py <- megadetector("/mnt/machinelearning/megaDetector/md_v5a.0.0.pt")}
}
