% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{load_img_resize}
\alias{load_img_resize}
\title{Load and resize an image and return an image tensor.}
\usage{
load_img_resize(file, height = 299, width = 299, standardize = TRUE)
}
\arguments{
\item{file}{path to a JPEG file}

\item{height}{the height the cropped image will be resized to.}

\item{width}{the width the cropped image will be resized to.}

\item{standardize}{standardize the image, TRUE or FALSE.}
}
\value{
An image tensor.
}
\description{
Load and resize an image and return an image tensor. Internal function to be called by image generator function.
}
