% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{load_img_resize_crop}
\alias{load_img_resize_crop}
\title{Load, resize and crop an image and return an image tensor.}
\usage{
load_img_resize_crop(data, height = 299, width = 299, standardize = TRUE)
}
\arguments{
\item{data}{a list with the first element being a path to an image file and the next four arguments being the bounding box coordinates.}

\item{height}{the height the cropped image will be resized to.}

\item{width}{the width the cropped image will be resized to.}

\item{standardize}{standardize the image, TRUE or FALSE.}

\item{batch_size}{the batch size for the image generator.}
}
\value{
A Tensorflow image data generator.
}
\description{
Load a JPEG image and crop it to a bounding box. Internal function to be called by image generator function.
}
