% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{load_yaml}
\alias{load_yaml}
\title{Load data from a YAML file.}
\usage{
load_yaml(file)
}
\arguments{
\item{file}{the full path of the file to load}
}
\value{
data extracted from the file, dict form
}
\description{
Load data from a YAML file.
}
\examples{
\dontrun{
config <- load_yaml('config.yml')
}
}
