% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMDModel.R
\name{loadMDModel}
\alias{loadMDModel}
\title{Load MegaDetector model file from directory or file}
\usage{
loadMDModel(modelfile)
}
\arguments{
\item{modelfile}{.pb file or directory obtained from megaDetector}
}
\value{
a tfsession containing the MD model
}
\description{
Load MegaDetector model file from directory or file
}
\examples{
\dontrun{
mdmodel <- "megadetector_v4.1.pb"
mdsession <- loadMDModel(mdmodel)
}
}
