% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{loadModel}
\alias{loadModel}
\title{Title}
\usage{
loadModel(model_path, class_file, device = "gpu")
}
\arguments{
\item{model_path}{path to model}

\item{class_file}{path to class list}

\item{device}{send model to the specified device}
}
\description{
Title
}
