% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMD.R
\name{parseMD}
\alias{parseMD}
\title{parse MD JSON results file into a simple dataframe}
\usage{
parseMD(allframes, mdresults)
}
\arguments{
\item{allframes}{dataframe containing all frames}

\item{mdresults}{raw MegaDetector output}
}
\value{
original dataframe including md results
}
\description{
parse MD JSON results file into a simple dataframe
}
\examples{
\dontrun{
mdresults <- parseMDsimple(mdres)
}
}
