% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMDjson.R
\name{parseMDjson}
\alias{parseMDjson}
\title{converte the JSON file produced bye the
Python version of MegaDetector into the format
produced by detectObjectBatch}
\usage{
parseMDjson(json)
}
\arguments{
\item{json}{json data in a list format}
}
\value{
a list of MegaDetector results
}
\description{
converte the JSON file produced bye the
Python version of MegaDetector into the format
produced by detectObjectBatch
}
\examples{
\dontrun{
mdresults <- parseMDjson(json)
}
}
