% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMDjson.R
\name{parseMDsimple}
\alias{parseMDsimple}
\title{parse MD JSON results file into a simple dataframe}
\usage{
parseMDsimple(imagesall, mdresults)
}
\arguments{
\item{imagesall}{dataframe containing all frames}

\item{mdresults}{raw MegaDetector output}
}
\value{
original dataframe including md results
}
\description{
parse MD JSON results file into a simple dataframe
}
\examples{
\dontrun{
mdresults <- parseMDsimple(mdres)
}
}
