% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBoxes.R
\name{plotBoxesFromFlat}
\alias{plotBoxesFromFlat}
\title{Plot bounding boxes from MD flat data frame}
\usage{
plotBoxesFromFlat(image, label = FALSE, minconf = 0)
}
\arguments{
\item{image}{The mdres for the image}

\item{label}{T/F toggle to plot MD category}

\item{minconf}{minimum confidence to plot box}
}
\value{

}
\description{
Plot bounding boxes from MD flat data frame
}
