% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{predictSpecies}
\alias{predictSpecies}
\title{Title}
\usage{
predictSpecies(
  detections,
  model,
  classes,
  device = "gpu",
  out_file = NULL,
  file_col = "Frame",
  crop = TRUE,
  resize = 299,
  standardize = TRUE,
  batch_size = 1,
  workers = 1,
  channel_last = FALSE,
  raw = FALSE
)
}
\arguments{
\item{detections}{}

\item{model}{}

\item{classes}{}

\item{device}{}

\item{out_file}{}

\item{file_col}{}

\item{crop}{}

\item{resize}{}

\item{standardize}{}

\item{batch_size}{}

\item{workers}{}

\item{channel_last}{}

\item{raw}{}
}
\description{
Title
}
