% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectMD.R
\name{processYOLO5}
\alias{processYOLO5}
\title{Process YOLO5 output and convert to MD format}
\usage{
processYOLO5(n, boxes, classes, scores, selection, batch)
}
\arguments{
\item{n}{index for the record in the batch}

\item{boxes}{array of boxes returned by combined_non_max_suppression}

\item{classes}{vector of classes returned by combined_non_max_suppression}

\item{scores}{vector of probabilities returned by combined_non_max_suppression}

\item{selection}{vector of number of detected boxes returned by combined_non_max_suppression}

\item{batch}{batch used to detect objects}
}
\value{
a list of MD bounding boxes, classes, and confidence for the image
}
\description{
Returns a list with the standard MD output format. Used for batch processing
}
