% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reid.R
\name{remove_diagonal}
\alias{remove_diagonal}
\title{Removes the diagonal elements from a square matrix.}
\usage{
remove_diagonal(A)
}
\arguments{
\item{A}{square matrix}
}
\value{
martix A with diagonals removed
}
\description{
Removes the diagonal elements from a square matrix.
}
\examples{
\dontrun{cleaned_dist <- remove_diagonal(dist_matrix)}
}
