% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBoxes.R
\name{resizePad}
\alias{resizePad}
\title{Resize an image with padding}
\usage{
resizePad(img, size = 256)
}
\arguments{
\item{img}{the image, read by jpeg library}

\item{size}{new size}
}
\value{
returns resized jpeg image
}
\description{
Resize an image with padding
}
\examples{
\dontrun{
crop <- resizePad(cropped_image_path,256)
}
}
