% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{save_json}
\alias{save_json}
\title{Save data to a JSON file.}
\usage{
save_json(data, out_file, prompt = TRUE)
}
\arguments{
\item{data}{the dictionary to be saved}

\item{out_file}{full path to save file to}

\item{prompt}{prompt user to confirm overwrite}
}
\value{
None
}
\description{
Save data to a JSON file.
}
\examples{
\dontrun{
save_json(mdresults, 'mdraw.json')
}
}
