% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setEmpty.R
\name{setEmpty}
\alias{setEmpty}
\title{Title}
\usage{
setEmpty(allframes)
}
\arguments{
\item{allframes}{all megadetector allframes}
}
\value{
list of empty/human/vehicle allframes with md classification
}
\description{
Title
}
\examples{
\dontrun{
setEmpty(imagesall)
}
}
