% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setLocation.R
\name{setLocation}
\alias{setLocation}
\title{Set Region/Site/Camera Information for Each Image}
\usage{
setLocation(
  files,
  basedir,
  adjust = 0,
  rename = TRUE,
  region = NA,
  site = NA,
  camera = NA
)
}
\arguments{
\item{files}{file manifest data frame}

\item{basedir}{directory from which files are obtained}

\item{adjust}{adjust starting position in folder hierarchy, defaults to 0}

\item{rename}{create a new unique name from region/site/camera/date, defaults to true}

\item{region}{general region of camera deployment, can be inputted manually or pulled from directory}

\item{site}{site of camera deployment, can be inputted manually or pulled from directory}

\item{camera}{camera name, can be inputted manually or pulled from directory}
}
\value{
file manifest with additional columns for region/site/camera
}
\description{
Set Region/Site/Camera Information for Each Image
}
\examples{
\dontrun{
setLocation(files,basedir)
}
}
