% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupDirectory.R
\name{setupDirectory}
\alias{setupDirectory}
\title{Title}
\usage{
setupDirectory(imagedir)
}
\arguments{
\item{imagedir}{local directory that contains data to process}
}
\value{
None
}
\description{
Title
}
\examples{
\dontrun{
setupDirectory(/home/kyra/animl/examples)
}
}
