% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupEnvironment.R
\name{setupEnv}
\alias{setupEnv}
\title{Setup Conda Environment}
\usage{
setupEnv(config = "animlenv.yml")
}
\arguments{
\item{config}{environment config yaml file}
}
\value{
none
}
\description{
Setup Conda Environment
}
\examples{
\dontrun{
setupEnv()
}
}
