% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{single_classification}
\alias{single_classification}
\title{Get Maximum likelihood label for each Detection}
\usage{
single_classification(animals, predictions_raw, class_list)
}
\arguments{
\item{predictions_raw}{softmaxed likelihoods from predict_species}

\item{class_list}{list of class labels}

\item{detections}{manifest of animal detections}
}
\value{
dataframe with prediction and confidence columns
}
\description{
Get Maximum likelihood label for each Detection
}
\examples{
\dontrun{animals <- single_classification(animals, pred_raw, class_list)}
}
