% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symlinkClassification.R
\name{symlinkClasses}
\alias{symlinkClasses}
\title{Create SymLink Directories and Sort Classified Images}
\usage{
symlinkClasses(alldata, linkdir)
}
\arguments{
\item{alldata}{DataFrame of classified images}

\item{linkdir}{Destination directory for symlinks}
}
\value{
none
}
\description{
Create SymLink Directories and Sort Classified Images
}
\examples{
\dontrun{}
}
