% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symlink.R
\name{symlinkMD}
\alias{symlinkMD}
\title{Create SymLink Directories and Sort Classified Images Based on MD Results}
\usage{
symlinkMD(manifest, linkdir, outfile = NULL, copy = FALSE)
}
\arguments{
\item{manifest}{DataFrame of classified images}

\item{linkdir}{Destination directory for symlinks}

\item{outfile}{Results file to save to}

\item{copy}{Toggle to determine copy or hard link, defaults to link}
}
\value{
manifest with added link columns
}
\description{
Create SymLink Directories and Sort Classified Images Based on MD Results
}
\examples{
\dontrun{
symlinkMD(manifest, linkdir)
}
}
