% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{test_classifier}
\alias{test_classifier}
\title{Test a model with a Config file}
\usage{
test_classifier(cfg)
}
\arguments{
\item{cfg}{config .yml file containing test settings}
}
\value{
None
}
\description{
Test a model with a Config file
}
\examples{
\dontrun{test_classifier('training_cfg.yml')}
}
