% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testMD.R
\name{testMD}
\alias{testMD}
\title{Title}
\usage{
testMD(imagesall, mdsession)
}
\arguments{
\item{imagesall}{dataframe of all images}

\item{mdsession}{MegaDetector tensorflow session}
}
\value{
Null, plots box on image
}
\description{
Title
}
\examples{
\dontrun{
testMD(imagesall,mdsession)
}
}
