% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symlink.R
\name{unlink}
\alias{unlink}
\title{Remove Symlinks}
\usage{
unlink(manifest)
}
\arguments{
\item{manifest}{DataFrame of classified images}
}
\value{
manifest without link column
}
\description{
Remove Symlinks
}
\examples{
\dontrun{
symlinkMD(manifest, linkdir)
}
}
