% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{update_animl_py}
\alias{update_animl_py}
\title{Update animl-py version}
\usage{
update_animl_py(py_env = "animl_env", animl_version = ANIML_VERSION)
}
\arguments{
\item{py_env}{name of python environment}

\item{animl_version}{version of animl to install}
}
\value{
None
}
\description{
Update animl-py version
}
\examples{
\dontrun{update_animl_py(py_env = "animl_env", animl_version = ANIML_VERSION)}
}
