% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{update_labels}
\alias{update_labels}
\title{Udate Results from File Browser}
\usage{
update_labels(manifest, link_dir, unique_name = "UniqueName")
}
\arguments{
\item{manifest}{dataframe containing file data and predictions}

\item{link_dir}{directory to sort files into}

\item{unique_name}{column name indicating a unique file name for each row}
}
\value{
dataframe with new "Species" column that contains the verified species
}
\description{
Udate Results from File Browser
}
\examples{
\dontrun{
results <- updateResults(resultsfile, linkdir)
}
}
