% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateResults.R
\name{updateResults}
\alias{updateResults}
\title{Udate Results from File Browser}
\usage{
updateResults(resultsfile, linkdir)
}
\arguments{
\item{resultsfile}{final results file with predictions, expects a "UniqueName" column}

\item{linkdir}{symlink directory that has been validated}
}
\value{
dataframe with new "Species" column that contains the verified species
}
\description{
Udate Results from File Browser
}
\examples{
\dontrun{
results <- updateResults(resultsfile, linkdir)
}
}
