% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{WorkingDirectory}
\alias{WorkingDirectory}
\title{Set Working Directory and Save File Global Variables}
\usage{
WorkingDirectory(workingdir, pkg.env)
}
\arguments{
\item{workingdir}{local directory that contains data to process}

\item{pkg.env}{environment to create global variables in}
}
\value{
None
}
\description{
Set Working Directory and Save File Global Variables
}
\examples{
\dontrun{
WorkingDirectory("/home/kyra/animl/examples",globalenv())
}
}
