% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{active_times}
\alias{active_times}
\title{Get start and stop dates for each camera folder.}
\usage{
active_times(
  manifest_dir,
  filepath = "filepath",
  camera_depth = 0,
  timestamp_col = "datetime"
)
}
\arguments{
\item{manifest_dir}{either file manifest or directory of files to analyze}

\item{camera_depth}{directory depth from which to split cameras}

\item{timestamp_col}{column in manifest to use for datetime information, defaults to "datetime"}

\item{file_col}{column in manifest to use for file paths, defaults to "filepath"}
}
\value{
times dataframe with min and max timestamp per camera
}
\description{
Get start and stop dates for each camera folder.
}
\examples{
\dontrun{
active_times('path/to/data', recursive=TRUE, camera_depth=2)
}
}
