% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{animl_install}
\alias{animl_install}
\title{Load animl-py if available}
\usage{
animl_install(envname = "animl_env", python_version = "3.12")
}
\arguments{
\item{envname}{name of python environment}

\item{python_version}{version of python to install}
}
\value{
animl-py module
}
\description{
Load animl-py if available
}
\examples{
\dontrun{animl_py <- load_animl_py()}
}
