% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{animl_install}
\alias{animl_install}
\title{Load animl-py if available}
\usage{
animl_install(
  py_env = "animl_env",
  animl_version = ANIML_VERSION,
  python_version = "3.12",
  confirm = TRUE
)
}
\value{
animl-py module
}
\description{
Load animl-py if available
}
\examples{
\dontrun{animl_py <- load_animl_py()}
}
