% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{animl_install}
\alias{animl_install}
\title{Install animl-py Python environment}
\usage{
animl_install(envname = "animl_env", python_version = "3.12")
}
\arguments{
\item{envname}{name of python environment}

\item{python_version}{version of python to install}
}
\value{
invisible NULL
}
\description{
Install animl-py Python environment
}
\examples{
\dontrun{animl_install("animl_env")}
}
