% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyPredictions.R
\name{applyPredictions}
\alias{applyPredictions}
\title{Apply Classifier Predictions and Merge DataFrames}
\usage{
applyPredictions(
  animals,
  empty,
  classfile,
  pred,
  outfile = NULL,
  counts = FALSE
)
}
\arguments{
\item{animals}{Set of animal crops}

\item{empty}{Set of empty, vehicle and human crops}

\item{classfile}{.txt file containing common names for species classes}

\item{pred}{Classifier predictions for animal crops}

\item{outfile}{File to which results are saved}

\item{counts}{Returns a table of all predictions, defaults to FALSE}
}
\value{
fully merged dataframe with Species predictions
}
\description{
Apply Classifier Predictions and Merge DataFrames
}
\examples{
\dontrun{
alldata <- applyPredictions(animals,empty,classfile,pred,counts = FALSE)
}
}
