% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{build_file_manifest}
\alias{build_file_manifest}
\title{File Management Module}
\usage{
build_file_manifest(
  image_dir,
  exif = TRUE,
  out_file = NULL,
  data_timezone = NULL,
  station_depth = NULL,
  camera_depth = NULL,
  recursive = TRUE
)
}
\arguments{
\item{image_dir}{folder to search through and find media files}

\item{exif}{returns date and time information from exif data, defaults to true}

\item{out_file}{.csv file to save manifest as}

\item{data_timezone}{timezone in which data was collected}

\item{station_depth}{integer value indicating folder depth from root image_dir that contains station name}

\item{camera_depth}{integer value indicating folder depth from root image_dir that contains camera name}

\item{recursive}{Should directories be scanned recursively? Default TRUE}
}
\value{
files dataframe with or without file dates
}
\description{
This module provides functions and classes for managing files and directories.
}
\details{
Kyra Swanson 2023
Find Image/Video Files and Gather exif Data
}
\examples{
\dontrun{
files <- build_file_manifest("C:\\\\Users\\\\usr\\\\Pictures\\\\")
}
}
