% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileManifest.R
\name{buildFileManifest}
\alias{buildFileManifest}
\title{Extract exif data and create dataframe, adjust time if necessary.}
\usage{
buildFileManifest(imagedir, outfile = NULL)
}
\arguments{
\item{imagedir}{file path}

\item{outfile}{file path to which the data frame should be saved}
}
\value{
files dataframe
}
\description{
Extract exif data and create dataframe, adjust time if necessary.
}
\examples{
\dontrun{
files <- extractFiles("C:\\\\Users\\\\usr\\\\Pictures\\\\")
}
}
