% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileManifest.R
\name{buildFileManifest}
\alias{buildFileManifest}
\title{Extract exif Data and Create File Manifest}
\usage{
buildFileManifest(imagedir, exif = TRUE, offset = 0, outfile = NULL)
}
\arguments{
\item{imagedir}{file path}

\item{exif}{returns date and time information from exif data, defaults to true}

\item{offset}{add offset to videos, defaults to 0}

\item{outfile}{file path to which the data frame should be saved}
}
\value{
files dataframe \link{FilePath,FileModifyDate}
}
\description{
Extract exif Data and Create File Manifest
}
\examples{
\dontrun{
files <- extractFiles("C:\\\\Users\\\\usr\\\\Pictures\\\\")
}
}
