% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileManagement.R
\name{buildFileManifest}
\alias{buildFileManifest}
\title{Extract exif Data and Create File Manifest}
\usage{
buildFileManifest(
  imagedir,
  exif = TRUE,
  offset = 0,
  recursive = TRUE,
  outfile = NULL
)
}
\arguments{
\item{imagedir}{path to a directory with image or video files.}

\item{exif}{returns date and time information from exif data, defaults to true}

\item{offset}{add offset in hours for videos when using the File Modified date, defaults to 0}

\item{recursive}{Should directories be scanned recursively? Default TRUE}

\item{outfile}{file path to which the data frame should be saved}
}
\value{
files dataframe with or without file dates
}
\description{
Extract exif Data and Create File Manifest
}
\examples{
\dontrun{
files <- extractFiles("C:\\\\Users\\\\usr\\\\Pictures\\\\")
}
}
