% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{check_python}
\alias{check_python}
\title{Check that the python version is compatible with the current version of animl-py}
\usage{
check_python(python_version = "3.12", initialize = TRUE)
}
\arguments{
\item{python_version}{version of python to install}

\item{initialize}{load reticulate library if true}
}
\value{
none
}
\description{
Check that the python version is compatible with the current version of animl-py
}
\examples{
\dontrun{check_python(initialize=FALSE)}
}
