% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifySpecies.R
\name{classifySpecies}
\alias{classifySpecies}
\title{Classifies Crops Using Specified Models}
\usage{
classifySpecies(
  input,
  model,
  resize = 299,
  standardize = TRUE,
  batch = 32,
  workers = 1
)
}
\arguments{
\item{input}{either dataframe with MD crops or list of filenames}

\item{model}{models with which to classify species}

\item{resize}{resize images before classification, defaults to 299x299px}

\item{standardize}{standardize images, defaults to FALSE}

\item{workers}{number of cores}

\item{batch_size}{number of images processed in each batch (keep small)}
}
\value{
a matrix of likelihoods for each class for each image
}
\description{
Classifies Crops Using Specified Models
}
\examples{
\dontrun{
pred <- classifySpecies(imagesallanimal, paste0(modelfile, ".h5"),
                      resize = 456, standardize = FALSE, batch_size = 64, workers = 8)
}
}
