% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{create_pyenv}
\alias{create_pyenv}
\title{Install python if necessary and create the environment animl_env}
\usage{
create_pyenv(envname = "animl_env", python_version = "3.12")
}
\arguments{
\item{envname}{name of the virtual environment to create / use (default "animl_env")}

\item{python_version}{python version to add to environment}
}
\value{
invisible TRUE on success, otherwise stops or returns FALSE invisibly on failure
}
\description{
Install python if necessary and create the environment animl_env
}
