% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detect}
\alias{detect}
\title{Apply MegaDetector to a Given Batch of Images}
\usage{
detect(
  detector,
  image_file_names,
  resize_width,
  resize_height,
  letterbox = TRUE,
  confidence_threshold = 0.1,
  file_col = "Frame",
  batch_size = 1,
  num_workers = 1,
  device = NULL,
  checkpoint_path = NULL,
  checkpoint_frequency = -1
)
}
\arguments{
\item{detector}{preloaded md model}

\item{image_file_names}{list of image filenames, a single image filename, or folder}

\item{letterbox}{}

\item{confidence_threshold}{only detections above this threshold are returned}

\item{file_col}{select which column if image_file_names is a manifest}

\item{checkpoint_path}{path to checkpoint file}

\item{checkpoint_frequency}{write results to checkpoint file every N images}

\item{quiet}{print debugging statements when false, defaults to true}

\item{image_size}{overrides default image size, 1280}
}
\value{
list of dictionaries of MegaDetector detections
}
\description{
Apply MegaDetector to a Given Batch of Images
}
\examples{
\dontrun{mdres <- detectMD_batch(md_py, allframes$Frame)}
}
