% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_coco}
\alias{export_coco}
\title{Converts the .csv file to a COCO-formatted .json file.}
\usage{
export_coco(manifest, class_list, out_file, info = NULL, licenses = NULL)
}
\arguments{
\item{manifest}{dataframe containing images and associated detections}

\item{class_list}{dataframe containing class names and their corresponding IDs}

\item{out_file}{path to save the formatted file}

\item{info}{info section of COCO file, named list}

\item{licenses}{licenses section of COCO file, array}
}
\value{
coco formated json
}
\description{
Converts the .csv file to a COCO-formatted .json file.
}
\examples{
\dontrun{export_megadetector(manifest, output_file= 'results.json', detector='MDv6')}
}
