% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_coco}
\alias{export_coco}
\title{Export a manifest to COCO format.}
\usage{
export_coco(manifest, class_list, out_file, info = NULL, licenses = NULL)
}
\arguments{
\item{manifest}{dataframe containing images and associated predictions}

\item{class_list}{dataframe containing class names and their corresponding IDs}

\item{out_file}{path to save the COCO formatted file}

\item{info}{optional info section of COCO file}

\item{licenses}{optional licenses section of COCO file}
}
\value{
coco formatted json file saved to out_file
}
\description{
Export a manifest to COCO format.
}
\examples{
\dontrun{
export_coco(manifest, classes, "path/to/out.json")
}
}
