% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_folders}
\alias{export_folders}
\title{Create SymLink Directories and Sort Classified Images}
\usage{
export_folders(
  manifest,
  out_dir,
  out_file = NULL,
  label_col = "prediction",
  file_col = "filepath",
  timestamp_col = "datetime",
  station_col = "station",
  unique_name = "uniquename",
  copy = FALSE
)
}
\arguments{
\item{manifest}{DataFrame of classified images}

\item{out_dir}{Destination directory for species folders}

\item{out_file}{if provided, save the manifest to this file}

\item{label_col}{specify 'prediction' for species or 'category' for megadetector class}

\item{file_col}{Colun containing file paths}

\item{timestamp_col}{column containing timestamps in format "\%Y-\%m-\%d \%H:\%M:\%S"}

\item{station_col}{column containing station names}

\item{unique_name}{Unique image name identifier}

\item{copy}{Toggle to determine copy or hard link, defaults to link}
}
\value{
manifest with added link columns
}
\description{
Create SymLink Directories and Sort Classified Images
}
\examples{
\dontrun{
manifest <- export_folders(manifest, out_dir)
}
}
