% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBoxes.R
\name{extractBoxes}
\alias{extractBoxes}
\title{Extract bounding boxes for a single image and save as new images}
\usage{
extractBoxes(
  image,
  min_conf = 0,
  buffer = 2,
  plot = T,
  return.crops = F,
  save = F,
  resize = NA,
  outdir = "",
  quality = 0.8
)
}
\arguments{
\item{image}{single image, raw MD output}

\item{min_conf}{Confidence threshold (defaults to 0, not in use)}

\item{buffer}{Adds a buffer to the MD bounding box, defaults to 2px}

\item{plot}{T/F toggle to plot each crop in the plot window, defaults to True}

\item{return.crops}{T/F toggle to return list of cropped images, defaults to False}

\item{save}{T/F toggle to save output cropped, defaults to False}

\item{resize}{Size in pixels to resize cropped images, NA if images are not resized, defaults to NA}

\item{outdir}{Directory in which output cropped images will be saved}

\item{quality}{Compression level of output cropped image, defaults to 0.8}
}
\value{
a flattened dataframe containing crop information
}
\description{
Requires the unflattened raw MD output
}
\examples{

extractBoxes(mdresults)
}
