% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{get_empty}
\alias{get_empty}
\title{Return MD empty, vehicle and human images in a dataframe}
\usage{
get_empty(manifest)
}
\arguments{
\item{manifest}{all megadetector frames}
}
\value{
list of empty/human/vehicle allframes with md classification
}
\description{
Return MD empty, vehicle and human images in a dataframe
}
\examples{
\dontrun{
empty <- get_empty(imagesall)
}
}
