% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitData.R
\name{getAnimals}
\alias{getAnimals}
\title{Title}
\usage{
getAnimals(manifest)
}
\arguments{
\item{manifest}{all megadetector frames}
}
\value{
animal frames classified by MD
}
\description{
Title
}
\examples{
\dontrun{
animals <- getAnimals(imagesall)
}
}
