% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitData.R
\name{getEmpty}
\alias{getEmpty}
\title{Title}
\usage{
getEmpty(manifest)
}
\arguments{
\item{manifest}{all megadetector allframes}
}
\value{
list of empty/human/vehicle allframes with md classification
}
\description{
Title
}
\examples{
\dontrun{
setEmpty(imagesall)
}
}
