% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{imageAugmentationColor}
\alias{imageAugmentationColor}
\title{Perform image augmentation through random color adjustments on an image/label pair.}
\usage{
imageAugmentationColor(image, label, rng)
}
\arguments{
\item{image}{an image tensor.}

\item{label}{a label tensor.}

\item{rng}{a random number generator use to generate a random seed.}
}
\value{
An image and label tensor.
}
\description{
Performs image augmentation on a image/label pair for training. Uses random brightness,contrast,saturation, and hue.
}
