% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{load_animl}
\alias{load_animl}
\title{Load animl-py if available}
\usage{
load_animl(envname = "animl_env", python_version = "3.12", interactive = TRUE)
}
\arguments{
\item{envname}{name of python environment}

\item{python_version}{version of python to install}

\item{interactive}{bool, only message if interactive}
}
\value{
none
}
\description{
Load animl-py if available
}
\examples{
\dontrun{animl_install("animl_env", ANIML_VERSION, python_version="3.12")}
}
