% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{load_animl_py}
\alias{load_animl_py}
\title{Load animl-py if available}
\usage{
load_animl_py()
}
\value{
animl-py module
}
\description{
Load animl-py if available
}
\examples{
\dontrun{animl_py <- load_animl_py()}
}
