% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{load_classifier}
\alias{load_classifier}
\title{Load a Classifier Model with animl-py}
\usage{
load_classifier(model_path, len_classes, device = NULL, architecture = "CTL")
}
\arguments{
\item{model_path}{path to model}

\item{len_classes}{path to class list}

\item{device}{send model to the specified device}

\item{architecture}{model architecture}
}
\value{
list of c(classifier, class_list)
}
\description{
Load a Classifier Model with animl-py
}
\examples{
\dontrun{andes <- loadModel('andes_v1.pt','andes_classes.csv')}
}
