% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{load_classifier}
\alias{load_classifier}
\title{Load a Classifier Model and Class_list}
\usage{
load_classifier(model_path, classes, device = NULL, architecture = "CTL")
}
\arguments{
\item{model_path}{path to model}

\item{classes}{path to class list or loaded class list}

\item{device}{send model to the specified device}

\item{architecture}{model architecture}
}
\value{
classifier model, class list
}
\description{
Load a Classifier Model and Class_list
}
\examples{
\dontrun{
classes <- load_class_list('sdzwa_andes_v1_classes.csv')
andes <- load_classifier('andes_v1.pt', nrow(classes))}
}
