% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{load_classifier}
\alias{load_classifier}
\title{Load a Classifier Model and Class_list}
\usage{
load_classifier(
  model_path,
  classes,
  device = NULL,
  architecture = "efficientnet_v2_m",
  quiet = TRUE
)
}
\arguments{
\item{model_path}{path to model}

\item{classes}{path to class list or loaded class list}

\item{device}{send model to the specified device}

\item{architecture}{model architecture}

\item{quiet}{bool, provide device information to user}
}
\value{
list of: classifier model, class list, and optionally epoch if resuming training
}
\description{
Load a Classifier Model and Class_list
}
