% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{load_data}
\alias{load_data}
\title{Load .csv or .Rdata file}
\usage{
load_data(file)
}
\arguments{
\item{file}{the full path of the file to load}
}
\value{
data extracted from the file
}
\description{
Load .csv or .Rdata file
}
\examples{
\dontrun{
  loadData("path/to/newfile.csv")
}
}
