% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reid.R
\name{load_miew}
\alias{load_miew}
\title{Load MiewID model}
\usage{
load_miew(file_path, device = NULL)
}
\arguments{
\item{file_path}{path to model weights}

\item{device}{toggle cpu or gpu}
}
\value{
meiwid model
}
\description{
Load MiewID model
}
\examples{
\dontrun{miew = load_miewid("/home/kyra/matchypatchy/Models/miewid_v3.bin")}
}
