% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMDmodel.R
\name{loadMDModel}
\alias{loadMDModel}
\title{Title}
\usage{
loadMDModel(modelfile)
}
\arguments{
\item{modelfile}{.pb file obtained from megaDetector}
}
\value{
a tfsession containing the MD model
}
\description{
Title
}
\examples{
\dontrun{
mdmodel<-"megadetector_v4.1.pb"
mdsession<-loadMDModel(mdmodel)
}
}
