% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{loadModel}
\alias{loadModel}
\title{Load a Classifier Model with animl-py}
\usage{
loadModel(model_path, class_file, device = NULL)
}
\arguments{
\item{model_path}{path to model}

\item{class_file}{path to class list}

\item{device}{send model to the specified device}
}
\value{
list of c(classifier, class_list)
}
\description{
Load a Classifier Model with animl-py
}
\examples{
\dontrun{andes <- loadModel('andes_v1.pt','andes_classes.csv')}
}
