% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{megadetector}
\alias{megadetector}
\title{Load MegaDetector}
\usage{
megadetector(model_path, device = NULL)
}
\arguments{
\item{model_path}{path to MegaDetector model (v5)}

\item{device}{load model onto given device description}
}
\value{
megadetector object
}
\description{
Load MegaDetector
}
\examples{
\dontrun{md_py <- megadetector("/mnt/machinelearning/megaDetector/md_v5a.0.0.pt")}
}
