% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{parse_detections}
\alias{parse_detections}
\title{parse MD results into a simple dataframe}
\usage{
parse_detections(
  results,
  manifest = NULL,
  out_file = NULL,
  threshold = 0,
  file_col = "frame"
)
}
\arguments{
\item{results}{json output from megadetector}

\item{manifest}{dataframe containing all frames}

\item{out_file}{path to save dataframe}

\item{threshold}{confidence threshold to include bbox}

\item{file_col}{column in manifest that refers to file paths}
}
\value{
original dataframe including md results
}
\description{
parse MD results into a simple dataframe
}
\examples{
\dontrun{
mdresults <- parseMD(mdres)
}
}
