% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMD.R
\name{parseMD}
\alias{parseMD}
\title{parse MD JSON results file into a simple dataframe}
\usage{
parseMD(mdresults, manifest = NULL, outfile = NULL)
}
\arguments{
\item{mdresults}{raw MegaDetector output}

\item{manifest}{dataframe containing all frames}

\item{outfile}{file path to save dataframe to}
}
\value{
original dataframe including md results
}
\description{
parse MD JSON results file into a simple dataframe
}
\examples{
\dontrun{
mdresults <- parseMD(mdres)
}
}
