% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{parseMD}
\alias{parseMD}
\title{parse MD results into a simple dataframe}
\usage{
parseMD(mdresults, manifest = NULL, outfile = NULL, buffer = 0.02)
}
\arguments{
\item{mdresults}{raw MegaDetector output}

\item{manifest}{dataframe containing all frames}

\item{outfile}{file path to save dataframe to}

\item{buffer}{percentage buffer to move bbox away from image edge}
}
\value{
original dataframe including md results
}
\description{
parse MD results into a simple dataframe
}
\examples{
\dontrun{
mdresults <- parseMD(mdres)
}
}
