% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMDjson.R
\name{parseMDjson}
\alias{parseMDjson}
\title{parse MD JSON results file}
\usage{
parseMDjson(json)
}
\arguments{
\item{json}{md output file}
}
\value{
list of MD detections
}
\description{
parse MD JSON results file
}
\examples{
\dontrun{
mdres <- parseMDjson("MDresults.json")
}
}
