% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxes.R
\name{plot_all_bounding_boxes}
\alias{plot_all_bounding_boxes}
\title{Plot all bounding boxes in a manifest}
\usage{
plot_all_bounding_boxes(
  manifest,
  out_dir,
  file_col = "frame",
  min_conf = 0.1,
  prediction = FALSE
)
}
\arguments{
\item{manifest}{manifest of detections}

\item{out_dir}{Name of the output directory}

\item{file_col}{Column name containing file paths}

\item{min_conf}{Confidence threshold to plot the box}

\item{prediction}{flag determining whether prediction be printed alongside bounding box}
}
\value{
None
}
\description{
Plot all bounding boxes in a manifest
}
\examples{
\dontrun{plot_all_bounding_boxes(manifest, 'Plots/''')}
}
