% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxes.R
\name{plot_box}
\alias{plot_box}
\title{Plot bounding boxes on image from md results}
\usage{
plot_box(image, label = FALSE, minconf = 0)
}
\arguments{
\item{image}{The mdres for the image}

\item{label}{T/F toggle to plot MD category}

\item{minconf}{minimum confidence to plot box}
}
\value{
no return value, produces bounding box in plot panel
}
\description{
Plot bounding boxes on image from md results
}
\examples{
\dontrun{
mdres <- classifyImageMD(mdsession, images$FilePath[30000])
plotBoxes(mdres, minconf = 0.5)
}
}
