% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxes.R
\name{plot_box}
\alias{plot_box}
\title{Plot bounding boxes on image from md results}
\usage{
plot_box(rows, file_col = "filepath", min_conf = 0, prediction = FALSE)
}
\arguments{
\item{rows}{row or rows of images in which the bounding box will be plotted}

\item{file_col}{Column name containing file paths}

\item{min_conf}{minimum confidence to plot box}

\item{prediction}{If True, display the prediction label alongside the bounding box.}
}
\value{
no return value, produces bounding box in plot panel
}
\description{
Plot bounding boxes on image from md results
}
\examples{
\dontrun{
test_image <- classify(classifier_model, test_image, file_col='filepath')
plot_box(test_image, file_col='filepath', minconf = 0.5, prediction=TRUE)
}
}
