% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_from_file}
\alias{plot_from_file}
\title{Read a CSV manifest file and perform box plotting on the images.}
\usage{
plot_from_file(csv_file, out_dir, file_col = "filepath")
}
\arguments{
\item{csv_file}{Path to the CSV file.}

\item{out_dir}{Saved location  of boxed images output dir.}

\item{file_col}{Column name containing file paths.}
}
\value{
None
}
\description{
Read a CSV manifest file and perform box plotting on the images.
}
\examples{
\dontrun{plot_from_file('manifest.csv', 'Plots/''')}
}
