% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBoxes.R
\name{plotBoxesFromFlat}
\alias{plotBoxesFromFlat}
\title{Plot bounding boxes from MD flat data frame}
\usage{
plotBoxesFromFlat(image, label = FALSE, minconf = 0)
}
\arguments{
\item{image}{The mdres for the image}

\item{label}{Toggle to plot MD category, defaults to FALSE}

\item{minconf}{minimum confidence to plot box}
}
\value{
no return value, produces bounding box in plot panel
}
\description{
Plot bounding boxes from MD flat data frame
}
\examples{
\dontrun{
mdres<-classifyImageMD(mdsession,images$FilePath[30000])
mdresflat<-flattenBoxesMDSimple(mdres)
plotBoxesFromFlat(mdresflat,minconf = 0.5)
}
}
