% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{predictSpecies}
\alias{predictSpecies}
\title{Infer Species for Given Detections}
\usage{
predictSpecies(
  detections,
  model,
  classes,
  device = NULL,
  out_file = NULL,
  file_col = "Frame",
  crop = TRUE,
  resize = as.integer(299),
  standardize = TRUE,
  batch_size = 1,
  workers = 1,
  channel_last = FALSE,
  raw = FALSE
)
}
\arguments{
\item{detections}{manifest of animal detections}

\item{model}{loaded classifier model}

\item{classes}{data.frame of classes}

\item{device}{send model to the specified device}

\item{out_file}{path to csv to save results to}

\item{file_col}{column in manifest containing file paths}

\item{crop}{use bbox to crop images before feeding into model}

\item{resize}{model expected dimensions}

\item{standardize}{standardize color}

\item{batch_size}{batch size for generator}

\item{workers}{number of processes}

\item{channel_last}{change matrix to BxWxHxC}

\item{raw}{output raw logits in addition to manifest}
}
\value{
detection manifest with added prediction and confidence columns
}
\description{
Infer Species for Given Detections
}
\examples{
\dontrun{animals <- predictSpecies(animals, classifier[[1]], classifier[[2]], raw=FALSE)}
}
