% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize_pad.R
\name{resize_pad}
\alias{resize_pad}
\title{Resize an image with padding}
\usage{
resize_pad(img, size = 256)
}
\arguments{
\item{img}{the image, read by jpeg library}

\item{size}{new size}
}
\value{
returns resized jpeg image
}
\description{
Resize an image with padding
}
\examples{
\dontrun{
crop<-resize_pad(cropped_image_path,256)
}
}
