% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{save_classifier}
\alias{save_classifier}
\title{Save model state weights}
\usage{
save_classifier(
  model,
  out_dir,
  epoch,
  stats,
  optimizer = NULL,
  scheduler = NULL
)
}
\arguments{
\item{model}{}

\item{out_dir}{}

\item{epoch}{}

\item{stats}{}

\item{optimizer}{}

\item{scheduler}{}
}
\description{
Save model state weights
}
