% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{save_classifier}
\alias{save_classifier}
\title{Save model state weights}
\usage{
save_classifier(
  model,
  out_dir,
  epoch,
  stats,
  optimizer = NULL,
  scheduler = NULL
)
}
\arguments{
\item{model}{pytorch model}

\item{out_dir}{directory to save model to}

\item{epoch}{current training epoch}

\item{stats}{performance metrics of current epoch}

\item{optimizer}{pytorch optimizer (optional)}

\item{scheduler}{pytorch scheduler (optional)}
}
\value{
None
}
\description{
Save model state weights
}
\examples{
\dontrun{save_classifier(model, 'models/', 10, {'acc': 0.85})}
}
