% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManagement.R
\name{saveData}
\alias{saveData}
\title{Save Data to Given File}
\usage{
saveData(data, outfile)
}
\arguments{
\item{data}{the dataframe to be saved}

\item{outfile}{the full path of the saved file}
}
\value{
none
}
\description{
Save Data to Given File
}
\examples{
\dontrun{
 saveData(files,"path/to/newfile.csv")
}
}
