% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{sequence_calculation}
\alias{sequence_calculation}
\title{Calculate sequence from timestamps}
\usage{
sequence_calculation(
  manifest,
  station_col,
  sort_columns = NULL,
  file_col = "filepath",
  timestamp_col = "datetime",
  maxdiff = 60
)
}
\arguments{
\item{manifest}{dataframe of images with station and timestamp columns}

\item{station_col}{a column in the animals and empty data frame that indicates the camera or camera station}

\item{sort_columns}{list of columns to sort by before calculating sequences. Defaults to None, which sorts by station_col and timestamp_col.}

\item{file_col}{column name representing the file path. Defaults to "filepath".}

\item{timestamp_col}{column name representing the timestamp in format "\%Y-\%m-\%d \%H:\%M:\%S". Defaults to "datetime".}

\item{maxdiff}{max time difference in seconds between sequences, default = 60}
}
\value{
manifest with sequence column, with a unique number associated with each sequence
}
\description{
Calculate sequence from timestamps
}
\examples{
\dontrun{
manifest <- sequence_calculation(manifest,'station')
}
}
