% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitAnimals.R
\name{setEmpty}
\alias{setEmpty}
\title{Title}
\usage{
setEmpty(crops)
}
\arguments{
\item{crops}{all megadetector crops}
}
\value{
list of empty/human/vehicle crops with md classification
}
\description{
Title
}
