% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupDirectory.R
\name{setupDirectory}
\alias{setupDirectory}
\title{Set Working Directory and Save File Global Variables}
\usage{
setupDirectory(workingdir)
}
\arguments{
\item{workingdir}{local directory that contains data to process}
}
\value{
None
}
\description{
Set Working Directory and Save File Global Variables
}
\examples{
\dontrun{
setupDirectory(/home/kyra/animl/examples)
}
}
