% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symlinkClassification.R
\name{symlinkMD}
\alias{symlinkMD}
\title{Create SymLink Directories and Sort Classified Images Based on MD Results}
\usage{
symlinkMD(alldata, linkdir, outfile = resultsfile, copy = FALSE)
}
\arguments{
\item{alldata}{DataFrame of classified images}

\item{linkdir}{Destination directory for symlinks}

\item{outfile}{Results file to save to}

\item{copy}{Toggle to determine copy or hard link, defaults to link}
}
\value{
none
}
\description{
Create SymLink Directories and Sort Classified Images Based on MD Results
}
