% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{test_main}
\alias{test_main}
\title{Test a model with a Config file}
\usage{
test_main(cfg)
}
\arguments{
\item{cfg}{config .yml file containing test settings}
}
\value{
None
}
\description{
Test a model with a Config file
}
\examples{
\dontrun{test_main('training_cfg.yml')}
}
