% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testMD.R
\name{testMD}
\alias{testMD}
\title{Select a Random Image and Run Through MegaDetector}
\usage{
testMD(input, mdsession, mdversion = 5, minconf = 0)
}
\arguments{
\item{input}{dataframe of all images}

\item{mdsession}{MegaDetector mdsession}

\item{mdversion}{megadetector version, defaults to 5}

\item{minconf}{minimum confidence with which to draw boxes, defaults to 0}
}
\value{
Null, plots box on image
}
\description{
Select a Random Image and Run Through MegaDetector
}
\examples{
\dontrun{
testMD(input, mdsession)
}
}
