% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testMD.R
\name{testMD}
\alias{testMD}
\title{Select a Random Image and Run Through MegaDetector}
\usage{
testMD(imagesall, mdsession, minconf = 0.9, frame = NA)
}
\arguments{
\item{imagesall}{dataframe of all images}

\item{mdsession}{MegaDetector tensorflow session}
}
\value{
Null, plots box on image
}
\description{
Select a Random Image and Run Through MegaDetector
}
\examples{
\dontrun{
testMD(imagesall, mdsession)
}
}
