% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train_main}
\alias{train_main}
\title{Model Training}
\usage{
train_main(cfg)
}
\arguments{
\item{cfg}{config .yml file containing training settings}
}
\value{
None
}
\description{
This module provides functions for training and testing classifier models
}
\details{
Kyra Swanson 2025
Train a model with a Config file
}
\examples{
\dontrun{train_main('training_cfg.yml')}
}
