% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{train_val_test}
\alias{train_val_test}
\title{Splits the manifest into training validation and test datasets for training}
\usage{
train_val_test(
  manifest,
  out_dir = NULL,
  label_col = "class",
  file_col = "filepath",
  percentage = c(0.7, 0.2, 0.1),
  seed = NULL
)
}
\arguments{
\item{manifest}{list of files to split for training}

\item{out_dir}{location to save split lists to}

\item{label_col}{column name containing class labels}

\item{file_col}{column containing file paths}

\item{percentage}{fraction of data dedicated to train-val-test}

\item{seed}{RNG seed, if none will pick one at random}
}
\value{
train manifest, validate manifest, test manifest, stats file
}
\description{
Splits the manifest into training validation and test datasets for training
}
\examples{
\dontrun{
  output <- train_val_test(manifest)
}
}
