% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{update_animl_py}
\alias{update_animl_py}
\title{Update animl-py version for the given environment}
\usage{
update_animl_py(envname = "animl_env")
}
\arguments{
\item{envname}{name of python environment}
}
\value{
None
}
\description{
Update animl-py version for the given environment
}
\examples{
\dontrun{update_animl_py(py_env = "animl_env")}
}
